Language column to db added. Nickname is now not required. All primary keys that become from Telegram is long(BIGINT)

+++ eAdvertisement_bot/Models/DbEntities/Advertisement.cs
@@ -12,7 +12,7 @@ namespace eAdvertisement_bot.Models.DbEntities
     public class Advertisement
     {
         [Column("channel_id")]
-        public int Channel_Id { get; set; }
+        public long Channel_Id { get; set; }
         public Channel Channel { get; set; }
 
         [Column("date_time")]
